requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
pymongo==4.2.0
redis==4.4.0
neo4j==5.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7